+++ src/flake8_comprehensions.py
@@ -34,6 +34,7 @@ class ComprehensionChecker:
         "C410": "C410 Unnecessary {type} passed to list() - ",
         "C411": "C411 Unnecessary list call - remove the outer call to list().",
         "C412": "C412 Unnecessary list comprehension - 'in' can take a generator.",
+        "C413": "C413 Unnecessary {outer} call around {inner}().",
     }
 
     def run(self):
@@ -178,6 +179,49 @@ class ComprehensionChecker:
                         self.messages["C408"].format(type=node.func.id),
                         type(self),
                     )
+
+                elif (
+                    node.func.id in {"list", "reversed"}
+                    and num_positional_args > 0
+                    and isinstance(node.args[0], ast.Call)
+                    and isinstance(node.args[0].func, ast.Name)
+                    and node.args[0].func.id == "sorted"
+                ):
+                    msg = self.messages["C413"]
+                    if node.func.id == "reversed":
+                        reverse_value = (
+                            [False]
+                            if not node.args[0].keywords
+                            else [
+                                kw.arg == "reverse"
+                                and (
+                                    isinstance(kw.value, ast.NameConstant)
+                                    and kw.value.value
+                                    or isinstance(kw.value, ast.Num)
+                                    and kw.value.n != 0
+                                )
+                                for kw in node.args[0].keywords
+                                if isinstance(kw.value, (ast.NameConstant, ast.Num))
+                            ]
+                        )
+                        if reverse_value:
+                            msg = msg.rstrip(
+                                "."
+                            ) + " - use sorted(..., reverse={!r}).".format(
+                                not reverse_value[0]
+                            )
+                        else:
+                            msg = (
+                                msg.rstrip(".")
+                                + " - toggle reverse argument to sorted()."
+                            )
+                    yield (
+                        node.lineno,
+                        node.col_offset,
+                        msg.format(inner=node.args[0].func.id, outer=node.func.id),
+                        type(self),
+                    )
+
             elif isinstance(node, ast.Compare):
                 if (
                     len(node.ops) == 1